# opt: make lookup join more expensive