… pkg/sql/opt/norm/testdata/rules/combo
… pkg/sql/opt/xform/testdata/external/tpch