@knz 2019-03-08
Which process did you use to pick the constant? (Even if manual - would be good to know)