@justinj 2019-03-12
For completeness, this change doesn't change the plans for any of the private queries.